src/main/persistence/JSonWriter.java
package persistence;


import model.Question;
import model.Set;
import org.json.JSONArray;
import org.json.JSONObject;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.PrintWriter;

// Represents a writer that writes JSON representation of set to file
public class JSonWriter {
    private static final int TAB = 4;
    private String destination;
    private PrintWriter writer;


    // EFFECTS: constructs writer to write to destination file
    public JSonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of set to file
    public void write(Set set) {
        JSONObject json = set.toJSon();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: opens writer
    //          throws FileNotFoundException if destination file cannot be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
    }

}
